tests/sched_xetparam.c: print syscall() result using %ld format

+++ tests/sched_xetparam.c
@@ -11,16 +11,17 @@
 int
 main(void)
 {
-	struct sched_param *const param = tail_alloc(sizeof(struct sched_param));
-	int rc = syscall(__NR_sched_getparam, 0, param);
-	printf("sched_getparam(0, [%d]) = %d\n",
+	struct sched_param *const param =
+		tail_alloc(sizeof(struct sched_param));
+
+	long rc = syscall(__NR_sched_getparam, 0, param);
+	printf("sched_getparam(0, [%d]) = %ld\n",
 	       param->sched_priority, rc);
 
 	param->sched_priority = -1;
 	rc = syscall(__NR_sched_setparam, 0, param);
-	printf("sched_setparam(0, [%d]) = %d %s (%m)\n",
-	       param->sched_priority, rc,
-	       errno2name());
+	printf("sched_setparam(0, [%d]) = %ld %s (%m)\n",
+	       param->sched_priority, rc, errno2name());
 
 	puts("+++ exited with 0 +++");
 	return 0;
